terraform/shared/apt-repo/gpg-keys/requirements.txt
boto3
python-gnupg
